fix: labview: protected characters (;) in flatten to csv
VI Flatten to CSV Record.vi was not working properly if someone
used '; ' as delimiter - the space ' ' caused it protected only sequence
'; ' by quotes and not just ';'.

Also there is a new possiblity to add protected characters.

This is used in Create 1/2D Table that semicolon is always protected
Especially needed in Create 1D Table collumn vector, where simplified
method was used.